Fix connection failure when the initial mode was listen not ask.
The code was assuming the central manager state and peripheral manager state would always agree, but of course they don't.  So now the bluetooth manager keeps each state independently and just notifies a "composite" state to the main view.  This ensures we stay in the main view until we are ready both to listen and to whisper.