Tweak the display of multiple listeners.
This completes a first stabilization pass on the ability to handle multiple whisperers.

The ListenView now waits a bit when starting up (5 seconds) to see if it sees multiple whisperers in the environment.  If so, it requires the user to choose which whisperer they want to listen to.  This is a live popover display from the status bar, which changes to read "Tap to select a whisperer..."