Add optional speaking of whispered or listened text.
(This is the third Leanne change.)

Also fixed the regression of not waiting for bluetooth or offering to enable it.

Also added a settings button to the main view.

There are some bugs to work out:

- When the whisperer is speaking, it should also be playing alert sounds, but it's not.

- The sizing of the text in the controls should be immune to dynamic type size, but it's not.

There are also some tweaks needed:

The buttons on the front page need to be aligned in width so they center properly with respect to each other.