Start the refactor to introduce transports.
In order to run the app in the simulator, we need a transport that doesn't
require Bluetooth.  So we refactor the logic of the Bluetooth manager and
the Whisper ViewModel into three layers: the transport layer (Bluetooth Layer),
the transport (Bluetooth Whisper Transport), and the controller (Whisper View Model).
That refactor allows us to introduce an alternate "Dribble" layer which does
nothing but put the whispers into a file for later use (and, of course, works
well in the simulator).

This checkin introduces the bottom two layers and refactors the WhisperViewModel
to use the new layers.  This is runnable in Whisper mode in the simulator,
and simulates picking up two listeners, one at 2 seconds, and one at 4 seconds.

Other work includes anticipating a two-way communication layer based on Ably
pub/sub over websockets with the introduction of "control packets" into the
Text Protocol.