Start to cut in the combination view.
The combination view will handle sending to multiple remotes,
and making the viewmodel logic independent of the transport type.

In order to allow manual discovery and auto-discovery to work together, I added
a preference not to auto-select a solo discovered whisperer.

Also fixed a bug where the control of whether to show the WhisperersView
was held in the ListenView rather than the ListenViewModel.