Make the app usable without Bluetooth.
If Bluetooth is not available, the Listen button is disabled and a message is shown explaining why.  If neither Bluetooth nor internet is available, both Whisper and Listen are disabled and a message is shown explaining why.  If Bluetooth is available but internet is not, no message is shown but the "Share Listen Link" button is disabled when whispering.

This is build 42 (external), and will be submitted for app review.