Make whispering work when Bluetooth is off.
Now the combo whisper transport is sensitive to the status of both underlying transports, and will only use the ones that are working for whispering.  If one stop working in the middle of the session, the session will be closed with a message that says that.

Along with this, the content of the URL being supplied to the content transport is now
owned higher up in the app, to allow for conversation IDs to be used instead of the client ID.  So the logic provided by the combo transport to the choice view just adjusts the URL based on whether the internet transport is available for sending it or not.