Tweak Stop Listening action to fix disablement issue.
When the status view was popped up, the Stop Listening button in the control view was disabled, but pressing on it would trigger its action.  The action of popping up the alert would then fail (presumably because it was on top of the status detail view), but the presentation value was still true so there was no way to re-invoke the alert.

Changing the button action to a toggle at least allows recovery from this situation by hitting it twice, but this feels like an Apple bug.