Clean up how the user exits a current session.
Two changes were made, one to fix #46 and one to fix #45.  But they both had to be made together, because they interoperate:

For #46, the problem was that the stop alert was being popped on a sub-view, so if the main view had already come on top of the subview the alert wouldn't show.  So the solution was to move the alert handling into the superview and to give the subview a way to invoke it.

For #45, the problem was that the view could not be released while it had focus as long as there was a magic keyboard attached.  So we had to lose focus on the Whisper area before showing the dialog that confirms quitting.  That way, when the dialog was dismissed, there would be no focus and the view could exit.  (Because we remove focus before popping the alert, the "cancel" action in the alert was changed to restore focus to the view.) Since the alert handling was moved to the main window for #45, this change was easy to make.

Hopefully this will take care of both issues.