Don't drop Listener sessions when bluetooth network goes down.
There is no issue for this; it was a missed fix in the Combo Listener transport that was made a while ago in the Combo Whisper transport.  Since Bluetooth goes down every time the device sleeps, but then the connections come back intact when the device wakes up, we don't want to terminate sessions just because Bluetooth is unavailable.