Immediately drop Whisperer who refuses to listen to us.
We now show the dialog *after* we drop the connection.

This fixes #39 and relies on the fix of #47 to avoid a crash.

This also helps with the case where we have made a listen offer and received a refusal from the Whisperer.  Now we detect that and close the listen session with the same message we use when we get kicked out of an already-accepted conversation.