Add posting of changed conversation names and non-shared-profile changed usernames.
This is an enhancement to the fix for #41 - before we were posting just new conversations but now we are posting updates to conversation names and profile names.  This is all in service of clickonetwo/whisper-server#11 whose goal is to make sure listen links are always named correctly by the server.