Silently restart the TCP transport if it has an initialization error.
Testing seems to show that, if the Ably infrastructure is going to fail with the `NACK: count response is greater than the total of pending messages` error in a way that disables communications, it will fail when sending the very first packet.  So now the TCP whisper transport sends a dummy packet on startup, looks for an error, and if it encounters one it quickly does a restart of the transport layer (stop followed by start).

This change is what actually fixes #49. The "repeated packet" work of 5173208038 has been conditionalized out until real-world experience shows that the repeats are needed.