Have Bluetooth listeners exit when the Whisperer explicitly drops.
Experimentation shows that Bluetooth listeners see messages even when they are in the background or the device is sleeping, so they will see a drop when one occurs.  Since it's trivial to reconnect once dropped, and we've decided history is not important, there's no reason not to disconnect them when the Whisperer does.

This fixes #53, because the failure to reconnect only happens after an explicit disconnect (because the listeners think that whisperer has kicked them out).