Handle self-authorization for shared profiles.
We now recognize when a user is whispering to themself and automatically authorize them for the conversation (without adding it to the previous conversation list).  This allows people who use shared profiles to have one of their devices be dedicated to listening and have it always be able to listen to every conversation (including via local discovery) without ever being sent a link.

This fixes #64.

This was a deeper fix than I was expecting.  I was expecting the UI changes in the ListenProfileView to separate my conversations from those others have sent me links for, but I didn't realize I would have to auto-create a listen conversation from a whisper conversation on the fly.

While I was in the profile code, I also cleaned up logging around shared profile updates, put a 10 second throttle on profile updates (no more than one every 10 seconds), and restricted new updates to two situations (after the initial fetch at launch):

1. When a profile-sensitive view (such as choice view or whisper profile view or listen profile view) appears (or is uncovered by a modal going away)
2. When the app, in choice view (or one covering it), comes back to the front (i.e., becomes active)

This has greatly reduced traffic to the server but still has a good chance of noticing changes made on other machines even though the app is running.

Finally, I found a subtle bug on the server (clickonetwo/whisper-server#25) around shared profiles: since profile names are uploaded on launch, out-of-date shared clients could confuse the server by sending an out-of-date name for the shared profile.  This would in turn screw up other shared clients.

The fix was not to record launch-time names for shared profiles, only for local ones.  That way there's no chance of conflict and the updated name will still be known at next launch even if the original upload of it failed.