Introduce preferences to control single-tap action on Whisper and Listen.
We can't prefer Internet to Bluetooth unless we get rid of conversation discovery.  (Which is a good thing to do anyway, because discovery might make the app store think we are doing chat roulette.)  So now there are preferences for what single-tap does on the Whisper and Listen buttons.

For Whisper, the choices are to show the conversation window, start the default conversation, or start the most-recently used conversation.

For Listen, the choice are to show the conversation window or start the most-recently used conversation.

This is work on #68.