Work around cached if-none-match header behavior.
It was impossible to load a shared profile using profile ID and password because the If-None-Match headers from earlier calls were somehow being reused in the GET call to load the shared profile.  The workaround is to explicitly use If-None-Match headers with values that can't possibly match what's on the server.