Add a special scene for receiving listen links.
To properly support multiple windows, we need to open a new window for listen links that arrive while we are already listening or whispering.  This new scene does that, and closes when the listen session ends.

This, together with prior work, fixes #61 well enough to make multiple windows work cleanly.  But there are still additional fixes needed:

1. Definitely the background snapshots of listen and whisper sessions need to name them.
2. Maybe there should only ever be one scene that has the Whisper and Listen buttons?