Fix crash on launch from listen link.
It seems that, when the app is launched from a listen link, it may somehow launch a ListenView window with an empty conversation.  Not clear whether this was because the RootView wasn't yet established or because the ListenView was allowing for nil conversations.  But in any event making the ListenView be the RootView if it's given a nil conversation, and having the RootView not set the conversation to nil when the mode is .ask seems to have fixed the problem.