Don't auto-accept self except if profile is shared.
This fixes an edge case that's important for testing but is unlikely to appear in normal use.  If a user with a non-shared profile whispers to themself (using a second window) then they should have to accept themself and then see themselves listed in the Whisper and Listen profiles the same as with any other user.

Also remodularized an extension to a separate file.