Add versioning for settings.
This fixes #95.  Now, when a new build adds a preference, the old preference settings are not uploaded until a change is made to one of them, and the first build that does that will be pulled down by all the other (new) builds.  Older builds which don't understand the new preferences will ignore them until their build is updated, and any changes they make will be ignored by the server.