First changes to support server-side transcripts.
The TCP whisper transport now uses presence on the control channel to mark the start and end of conversations, so the server can use that as a trigger to start and end transcription of the conversation.  This is the first step needed to fix #92. Now what's needed is a way to retrieve those conversations from the client side.

There was also a fix made in the authenticator to better log server-side failures.