Prepare for protocol changes.
Replace fatalError calls on packets that aren't understood with anomaly logs, so that future protocol changes don't crash the app and the packets are just ignored.

Also replace every fatalError where ignoring request is possible into an anomaly so as to reduce the chance of crashes.  All the fatal errors that are left are "can't happen" situations where the crash site will reveal the exact problem and where the operation can't simply be ignored.